+++ apps/probability/calculation/finite_integral_calculation.h
@@ -12,15 +12,15 @@ public:
   int numberOfParameters() override;
   int numberOfEditableParameters() override;
   I18n::Message legendForParameterAtIndex(int index) override;
-  void setParameterAtIndex(float f, int index) override;
-  float parameterAtIndex(int index) override;
-  float lowerBound() override;
-  float upperBound() override;
+  void setParameterAtIndex(double f, int index) override;
+  double parameterAtIndex(int index) override;
+  double lowerBound() override;
+  double upperBound() override;
 private:
   void compute(int indexKnownElement) override;
-  float m_lowerBound;
-  float m_upperBound;
-  float m_result;
+  double m_lowerBound;
+  double m_upperBound;
+  double m_result;
 };
 
 }
